Entity Framework Core : Scaffold-DBContext - Part 2

+++ BookStoresWebAPI/BookStoresWebAPI/Models/BookAuthor.cs
@@ -8,7 +8,7 @@ namespace BookStoresWebAPI.Models
         public int AuthorId { get; set; }
         public int BookId { get; set; }
         public byte? AuthorOrder { get; set; }
-        public byte? RoyalityPercentage { get; set; }
+        public int? RoyalityPercentage { get; set; }
 
         public virtual Author Author { get; set; }
         public virtual Book Book { get; set; }
